requirements.txt
openai
pyyaml
pandas
numpy
